fix: compile MXFP4 scale alignment and isolate A5 test loading
Cast BYTE_BLOCK to the MMParams scale-stride type for Align template deduction. Disable FLA startup injection before Python, restore the private vendor after backend imports, and verify workspace/compute API ownership. Preserve the operator build during the wheel-only step.

Validation: reproduced the original scalar C++ compilation failure; corrected expression, 11 reference/metadata/compiler/launcher tests, and 4 storage tests pass locally. Full Ascend C compilation and A5 accuracy/performance remain pending.

+++ attention/mla_prolog/op_kernel/arch35/service_matmul_mxfp4_arch35.h
@@ -51,7 +51,7 @@ __aicore__ inline void MatmulMxFp4SplitK(
     mmLocalTensors<uint8_t, float> local;
     local.Init(buf);
     const uint32_t mAlign = Align(p.m, BLOCK_CUBE_SIZE);
-    const uint32_t scaleStrideA = paddedAScale ? Align(p.kScale, BYTE_BLOCK) : p.kScale;
+    const uint32_t scaleStrideA = paddedAScale ? Align(p.kScale, static_cast<uint32_t>(BYTE_BLOCK)) : p.kScale;
     const uint32_t cIdx = buf.cL0BufIter & 1u;
     auto cL0 = local.cL0Tensor[cIdx * L0C_PP_SIZE / sizeof(float)];
     WaitFlag<HardEvent::FIX_M>(L0C_EVENT0 + cIdx);
